add quick fix

+++ src/transformers/models/opt/modeling_opt.py
@@ -215,7 +215,7 @@ class OPTAttention(nn.Module):
             attn_weights = torch.max(attn_weights, torch.tensor(torch.finfo(attn_weights.dtype).min))
             attn_weights_before_softmax = attn_weights.view(bsz * self.num_heads, tgt_len, src_len)
 
-        attn_weights = nn.functional.softmax(attn_weights_before_softmax, dim=-1, dtype=torch.float32).to(attn_weights_before_softmax.dtype)
+        attn_weights = nn.functional.softmax(attn_weights_before_softmax, dim=-1, dtype=torch.float32)
 
         if layer_head_mask is not None:
             if layer_head_mask.size() != (self.num_heads,):
